Fixed issue with exporting to file only! - The script used called a database function, which was not initialized when only file export was selected.

+++ Main.py
@@ -69,7 +69,8 @@ Lesson scraping / Exporting function
 def export_all_tables(name, name_list, collection):
     print('- - - Scraping {} tables! - - - '.format(name))
 
-    Service_Connect.drop_collection(Database, collection)
+    if Settings.Database.Enabled:
+        Service_Connect.drop_collection(Database, collection)
 
     for table_name in name_list:
